Cart component route added
