+++ my-app/src/Cart.js
@@ -1,0 +1,11 @@
+
+
+function Cart({cartItems}){
+    return(
+        <>
+            <h1 className="cart-heading">Cart</h1>
+        </>
+    )
+}
+
+export default Cart;
